% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2h3.R
\name{s2h3}
\alias{s2h3}
\title{Get Corresponding S2 and H3 cells for given longitude and latitude}
\usage{
s2h3(data, long, lat, s2_level, h3_resolution)
}
\arguments{
\item{data}{The dataframe or tibble that contains a longitude and latitude field for individual points of interest}

\item{long}{The longitude field name in `data`}

\item{lat}{The latitude field name in `data`}

\item{s2_level}{The desired [s2 level](https://s2geometry.io/resources/s2cell_statistics.html)}

\item{h3_resolution}{The desired [h3 resolution](https://h3geo.org/docs/core-library/restable/)}
}
\description{
This function is a work in progress!
}
\examples{
sf_trees <- readr::read_csv('https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2020/2020-01-28/sf_trees.csv') \%>\%
filter(!is.na(latitude) | !is.na(longitude)) \%>\%
 s2h3(longitude,
      latitude,
      s2_level = 21,
      h3_resolution = 10)
}
